Cleaned up the homepage
- Moved the list and grid view widgets to a separate widget that handles the transition. This way the homepage widget now only handles the api calls and doesn't have any ui logic.
