+++ icapps_app/lib/Pages/HomePage/homePage.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:icapps_app/Api/beerService.dart';
 import 'package:icapps_app/Models/beer.dart';
-import 'package:icapps_app/Pages/HomePage/Widgets/viewTransition.dart';
-import 'package:icapps_app/Shared/modeIcon.dart';
+import 'package:icapps_app/Pages/HomePage/Widgets/beerOverview.dart';
 
 class HomePage extends StatefulWidget {
   HomePage({Key? key}) : super(key: key);
@@ -13,12 +12,6 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   late Future<List<Beer>> futureBeer;
-  int currentIndex = 0;
-  void changeCurrentViewMode(int newIndex) {
-    setState(() {
-      currentIndex = newIndex;
-    });
-  }
 
   @override
   void initState() {
@@ -35,26 +28,6 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        actions: [
-          ModeIcon(
-            icon: Icons.format_list_bulleted_rounded,
-            changeViewMode: changeCurrentViewMode,
-            index: 0,
-            selectedIndex: currentIndex,
-          ),
-          ModeIcon(
-            icon: Icons.grid_view_rounded,
-            changeViewMode: changeCurrentViewMode,
-            index: 1,
-            selectedIndex: currentIndex,
-          ),
-          SizedBox(
-            width: 20,
-          )
-        ],
-      ),
       body: Center(
           child: RefreshIndicator(
         backgroundColor: Theme.of(context).cardColor,
@@ -64,9 +37,8 @@ class _HomePageState extends State<HomePage> {
           future: futureBeer,
           builder: (BuildContext context, AsyncSnapshot<List<Beer>> data) {
             if (data.hasData) {
-              return Viewtransitioni(
+              return BeerOverview(
                 beers: data.data!,
-                currentIndex: currentIndex,
               );
             } else
               return Center(
